[RenderTarget2DArray] - fix up calculation of reported memory size to allow for >2GB
[at]jason.Nadro
#preflight 61494dc8315f5400013e890c
#lockdown jeff.farris

#ROBOMERGE-AUTHOR: andrew.firth
#ROBOMERGE-SOURCE: CL 17587810 via CL 17948984 via CL 18364628 via CL 18364725
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Staging -> Release-Engine-Test) (v895-18170469)

[CL 18364777 by andrew firth in ue5-release-engine-test branch]

+++ Engine/Source/Runtime/Engine/Private/TextureRenderTarget2DArray.cpp
@@ -78,12 +78,12 @@ void UTextureRenderTarget2DArray::GetResourceSizeEx(FResourceSizeEx& CumulativeR
 
 	// Calculate size based on format.
 	const EPixelFormat Format = GetFormat();
-	const int32 BlockSizeX	= GPixelFormats[Format].BlockSizeX;
-	const int32 BlockSizeY	= GPixelFormats[Format].BlockSizeY;
-	const int32 BlockBytes	= GPixelFormats[Format].BlockBytes;
-	const int32 NumBlocksX	= (SizeX + BlockSizeX - 1) / BlockSizeX;
-	const int32 NumBlocksY	= (SizeY + BlockSizeY - 1) / BlockSizeY;
-	const int32 NumBytes	= NumBlocksX * NumBlocksY * Slices * BlockBytes;
+	const int64 BlockSizeX	= GPixelFormats[Format].BlockSizeX;
+	const int64 BlockSizeY	= GPixelFormats[Format].BlockSizeY;
+	const int64 BlockBytes	= GPixelFormats[Format].BlockBytes;
+	const int64 NumBlocksX	= (SizeX + BlockSizeX - 1) / BlockSizeX;
+	const int64 NumBlocksY	= (SizeY + BlockSizeY - 1) / BlockSizeY;
+	const int64 NumBytes	= NumBlocksX * NumBlocksY * Slices * BlockBytes;
 
 	CumulativeResourceSize.AddUnknownMemoryBytes(NumBytes);
 }
